+++ src/mixins/DomHelper.ts
@@ -36,11 +36,47 @@ export class EventListener {
     }
 }
 
-export function componentFirstChild(children: HTMLElement[]): HTMLElement {
-    return children && children.filter(c => c && (c as GetNotificationOptions).tag)[0];
-}
-
 export function preventEventTarget(event: Event): void {
     event.preventDefault();
     event.stopPropagation();
+}
+
+export function isSVGElement(source?: HTMLElement | null): boolean {
+    if (!source) {
+        return false;
+    }
+
+    const idx = source.constructor.toString().search(/SVG.+Element/);
+    return idx > -1;
+}
+
+/**
+ * Check if `target` is indirect child of `parent`.
+ *
+ * @param {HTMLElement} parent The parent element.
+ * @param {HTMLElement} target The child element to check.
+ * @returns {boolean} TRUE if `target` is indirect child otherwise `false`.
+ */
+export function isChildOf(parent: HTMLElement | Node | null, target?: HTMLElement | null): boolean {
+    if (!target || !parent) {
+        return false;
+    }
+    if (parent.contains(target)) {
+        return true;
+    }
+
+    let result = false;
+    const children = (<HTMLElement>parent).children;
+
+    for (let i = 0; i < children.length; i++) {
+        const el = children.item(i);
+
+        if ((el?.id && target.id && (el?.id === target.id))
+            || (el === target) || isChildOf(el, target)) {
+            result = true;
+            return result;
+        }
+    }
+
+    return result;
 }